examples/python-pycairo/requirements.txt
pycairo==1.28.0
